+++ addon/components/flexberry-groupedit.js
@@ -361,6 +361,24 @@ export default FlexberryBaseComponent.extend({
   */
   tableStriped: true,
 
+  /**
+    Minimum column width, if width isn't defined in userSettings.
+
+    @property minAutoColumnWidth
+    @type Number
+    @default 150
+  */
+  minAutoColumnWidth: 150,
+
+  /**
+    Indicates whether or not invoke _setColumnWidths function on container resize.
+
+    @property widthChangeOnContainerResize
+    @type Boolean
+    @default true
+  */
+  widthChangeOnContainerResize: true,
+
   actions: {
     /**
       Handles action from object-list-view when no handler for this component is defined.